nuplayer: Allow devices to force a flush when surface changes
This fixes some playback problems on the Zenfone2 (running on 5.0
based blobs).

The problems are easily reproduced by viewing a video in Chrome and
then switching to full-screen mode.  The change in surface causes
the log to get flooded with messages like:

 dc_device_post: Buffer ID=14460 not imported by D

until Android crashes.  With this change, it ends up correctly
mapping the surfaces after the flush fixing the problem.

Change-Id: I105d647e724f5b63ef5cb0852481ecd2990c2d5e